### 2024-04-22 #6661 @editorialbot
[PRE REVIEW]: Simulation Decomposition in Python
**Submitting author:** @tupui (<a href="http://orcid.org/0000-0001-9816-1416">Pamphile ROY</a>)
**Repository:** https://github.com/Simulation-Decomposition/simdec-python
**Branch with paper.md** (empty if default branch): joss
**Version:** 1.1
