### 2024-04-22 #6661 @crvernon
:wave: @tupui - Thanks for your submission to JOSS.  It seems that your submission is closely related to one that was recently published in another software inclusive journal here:  https://www.sciencedirect.com/science/article/pii/S1364815223002840?via%3Dihub.  Thus, I have submitted this pre-review for editorial review so we may further evaluate this for uniqueness. This usually takes a week or two and I'll report back here as soon as I hear back from the rest of the editorial board.

Please feel free to comment here with any clarifying information that will help us make an informed decision.  Thanks!
