+++ joss/paper.bib
@@ -1,6 +1,5 @@
 @article{Herman2017,
   doi = {10.21105/joss.00097},
-  url = {https://doi.org/10.21105/joss.00097},
   year  = {2017},
   month = {jan},
   publisher = {The Open Journal},
@@ -18,15 +17,13 @@
   pages = {105898},
   year = {2024},
   issn = {1364-8152},
-  doi = {https://doi.org/10.1016/j.envsoft.2023.105898},
-  url = {https://www.sciencedirect.com/science/article/pii/S1364815223002840},
+  doi = {10.1016/j.envsoft.2023.105898},
   author = {Mariia Kozlova and Robert J. Moss and Julian Scott Yeomans and Jef Caers},
   keywords = {Global sensitivity analysis, Simulation decomposition, Monte Carlo simulation, Decision-making under uncertainty}
 }
 
 @article{Roy2023,
   doi = {10.21105/joss.05309},
-  url = {https://doi.org/10.21105/joss.05309},
   year = {2023},
   publisher = {The Open Journal},
   volume = {8},
@@ -46,5 +43,5 @@
   pages={261--272},
   year={2020},
   publisher={Nature Publishing Group},
-  doi={https://doi.org/10.1038/s41592-019-0686-2}
+  doi={10.1038/s41592-019-0686-2}
 }
