### 2024-04-22 #6661 @tupui
Hi @crvernon, thank you for starting up the process.

Yes the paper that you linked, and that we cite in our bibliography, is the main paper presenting the method itself. The paper was an in-depth presentation of the method itself and it was only noted that the method is made freely available.

The current submission is about the software itself and all the work that was done on making a Dashboard and making it accessible to everyone for free. This last part was never shared anywhere and entirely done after the first papers' submission.

Before someone brings up the authorship. The number of commits and lines that I wrote vs Mariia is not a relevant metric in this case. I could not have written any single line of code without Mariia's inputs and her own Matlab code which I ported. This was truly a joined collaboration, hence both our names are listed as co-authors.

@gnopik here is the start of the review. And feel free to add anything to what I said here.
