### 2024-04-24 2997062
Remove list inversions
