### 2024-04-30 647df87
Remove list inversions (#27)
Closes #26

There was some customs list inversion for the colour palette management
on the table side.
