+++ src/simdec/visualization.py
@@ -14,9 +14,9 @@ __all__ = ["visualization", "tableau", "palette"]
 
 
 SEQUENTIAL_PALETTES = [
+    "#26DCD1",
     "#DC267F",
     "#E8EA2F",
-    "#26DCD1",
     "#C552E4",
     "#3F45D0",
     "Oranges",
@@ -253,7 +253,7 @@ def tableau(
     table.insert(loc=0, column="N°", value=np.arange(1, stop=len(table) + 1)[::-1])
 
     # style the colour background with palette
-    cmap = mpl.colors.ListedColormap(palette[::-1])
+    cmap = mpl.colors.ListedColormap(palette)
     styler = table.style
     styler.format(precision=2)
     styler.background_gradient(subset=["colour"], cmap=cmap)
