### 2024-05-01 #6661 @crvernon
@tupui - While we are getting started, could you take a shot at fixing these:

> INVALID DOIs
>
> - https://doi.org/10.1016/j.envsoft.2023.105898 is INVALID because of 'https://doi.org/' prefix
> - https://doi.org/10.1038/s41592-019-0686-2 is INVALID because of 'https://doi.org/' prefix
