### 2024-05-01 #6661 @tupui
> Actually, I think I'll take this one on since you have been waiting a bit.

Thank you 🙇

> @tupui - While we are getting started, could you take a shot at fixing these:
>
> > INVALID DOIs
> >
> > * [doi.org/10.1016/j.envsoft.2023.105898](https://doi.org/10.1016/j.envsoft.2023.105898) is INVALID because of '[doi.org](https://doi.org/)' prefix
> > * [doi.org/10.1038/s41592-019-0686-2](https://doi.org/10.1038/s41592-019-0686-2) is INVALID because of '[doi.org](https://doi.org/)' prefix

I think I fixed it already and had relaunched the build of the proof. Something did not work? I don't have more ref of doi.org in the paper.
