### 2024-05-02 #6661 @crvernon
@JoshuaOsborneDATA would you be able to review this submission for JOSS per @shahmoradi's recommendation?
