### 2024-05-02 #6661 @matt-graham
> 👋 @matt-graham - Would you be willing to review this submission to JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html

Hi @crvernon. Yes I would be happy to review this submission - just as a heads up I wouldn't be able to start on review before 13th May.
