### 2024-05-12 da1f61a
Fix colour middle point
