### 2024-05-14 6659346
Limit EditableRangeSlider format to 3 decimals
