### 2024-05-18 a81bf18
Decompose with equal sample sized bins
