### 2024-05-18 08842c0
Remove unused cat code
