### 2024-05-21 da28df6
More robust min and add nudge to avoid unique values
