### 2024-05-22 Simulation-Decomposition/simdec-python#31 @matt-graham
Adding an example usage to documentation
Raising as part of JOSS review openjournals/joss-reviews/issues/6713

It would help users to have one or more example sin the documentation (or top-level README) of using `simdec` to compute sensitivity indices / visualize the results. This could potentially use the same data already used in the tests.
