### 2024-05-22 Simulation-Decomposition/simdec-python#32 @matt-graham
Missing parameter definition in docstring for `simdec.tableau`
Raising as part of JOSS review openjournals/joss-reviews/issues/6713

In the docstring for the `simdec.tableau` function

https://github.com/Simulation-Decomposition/simdec-python/blob/da28df6c981d5ac45f843a4b96fbfd227ed6604f/src/simdec/visualization.py#L191-L212

the argument `statistic` does not have a corresponding entry in parameter list in docstring.
