### 2024-05-22 #6713 @matt-graham
> I'm just checking in to see how the review is coming along. Could you please provide a short update in this thread. Thanks!

@crvernon I've started on review and raised some issues with regards to some points I've come across so far while working through checklist.

@tupui - a quick general question with regards to list of authors. Is `gnopik` the GitHub user for author 'Mariia Kozlova'? I'm checking as the reviewer checklist asks us to check that 'the full list of paper authors seem appropriate and complete' with some guidelines on what [constitutes authorship given in the JOSS documentation](https://joss.readthedocs.io/en/latest/submitting.html#authorship). I can see that `gnopik` has opened several issues and reviewed pull requests, which I believe would qualify for authorship under the JOSS guidelines, but as they do not have a name listed on their profile I wanted to check if this indeed the same person as included in the author list.
