### 2024-05-22 Simulation-Decomposition/simdec-python#34 @matt-graham
Reproducing figure in JOSS paper
Raising as part of JOSS review openjournals/joss-reviews/issues/6713

It would be nice if code was provided (even better if it was in the paper as an example usage) for reproducing the plot / table shown in Figure 1 in the paper, along with some description of the data used to produce it and how to access.
