### 2024-05-22 Simulation-Decomposition/simdec-python#35 @matt-graham
Suggested improvements and fixes to JOSS paper
Raising as part of JOSS review openjournals/joss-reviews/issues/6713

Some assorted changes I would recommend making to the JOSS paper

- [x] Define acronyms in full on first usage (AI, API, QMC)
- [x] Rephrase first paragraph in _Summary_ to make less disjointed
  - Perhaps something like 'Uncertainties are everywhere. Whether you are developing a new artificial intelligence system, running complex simulations or performing an experiment in a laboratory, uncertainties influence the system, and an approach is needed to understand how these uncertainties impact your results.' though maybe I'm not understanding the original text correctly. The connection of the 'And you need a ...' sentence to the preceding sentence is currently unclear as is what 'these' in the sentence refers to.
- [x] Rephrase second paragraph in _Summary_
   - Currently the last clause in the second sentence in paragraph doesn't make grammatical sense with respect to the first clause. I think perhaps something like 'Thanks to a clear Python API and an interactive web dashboard, [it / SimDec] makes uncertainty analysis accessible to everyone' would be better. I would potentially also scale back the claim that SimDec makes uncertainty analysis 'accessible to everyone' - which is quite a bold claim and difficult to verify, with something a bit more measured, for example 'SimDec offers an accessible approach to performing uncertainty analyses'.
- [x] Add bibliography entry and citation for 'AI Act'
  - Not clear what specifically is being referred to here. Remember the JOSS readership is international.
- [x] Add bibliography entry and citation for 'Better Regulation Guideline'
  - Again this needs a reference to make it clear what is being referred to.
- [x] Add bibliography entry and citation for Sobol' indices
  - https://en.wikipedia.org/wiki/Variance-based_sensitivity_analysis#References
- [x] Rephrase and expand 'decomposes the distribution of the output (target variable) by the multivariable scenarios'
  - This does not read right to me - the 'by' suggests the multivariable scenarios are the actor decomposing the distribution which I don't think is correct, though in general I think this sentence could do with expanding on to make it more understandable to a non-specialist audience (for example, what is a multivariable scenario here?)
- [x] Add more detail to caption of Figure 1 and make plot and table larger to improve readability
  - Currently it is difficult to understand what Figure 1 is showing as not much context is given. What is being plotted in the right panel and what is the table reporting? What 'input-output data' was used here. Aso the small size of the right panel makes it very difficult to read.
- [x] Correct typo 'through a Python packages available on PyPI' &rightarrow; 'through a Python package available on PyPI'
- [x] Make URL `https://simdec.io` pointing to web dashboard a hyperlink
- [ ] Expand on what 'powerful methods' are used from SALib
- [x] Correct spacing around hyphen between SciPy citations and 'notable the...' or make an em-dash &mdash;
- [x] Add bibliography entry and citation for Panel (https://zenodo.org/doi/10.5281/zenodo.3706648)
- [x] Add state of the field comparison to other comparable packages
  - For example `chaospy` allows computing sensitivity indices as does `SALib`, and it looks like the package [`sensitivity`](https://github.com/nickderobertis/sensitivity) might also be relevant.
