### 2024-05-22 #6713 @tupui
Hi @matt-graham, yes `gnopik` is Mariia Kozlova. I have written some text in the first JOSS issue regarding authorship, see this https://github.com/openjournals/joss-reviews/issues/6661#issuecomment-2071011136

Thank you for raising the other issues, I will reply there.
