### 2024-05-22 Simulation-Decomposition/simdec-python#34 @tupui
I would not do that in the paper. APIs tend to change and I would not want people to copy past something from a paper that we cannot easily update.

The figure was more of showcase not meant to be analysed by the reader, since the paper is less about the method and more about the code. i.e. we managed to make something pretty which is meaningful.

But if you really want, we can add a description.

Otherwise I would close this issue in favour #31 where you ask for an example in the doc itself, which we can run in the CI and update as we go.
