### 2024-05-23 Simulation-Decomposition/simdec-python#34 @matt-graham
In terms of APIs changing - while a good point, there are ways that could be dealt with, for example specifying the version of the package used if providing code in the paper or providing a permalink to a script in the repository at a specific commit. The reviewer checklist asks us to confirm that 'If the paper contains original results, results are entirely reproducible by reviewers' which at the moment I judged not to be the case, hence this issue.

I would say some form of description of what the figure is showing and the provenance of the data being visualized is a minimum requirement for the figure to be useful to a reader. At the moment from the current paper alone I struggled to understand what the figure was showing. Alternatively there is no requirement to include any figures so it could be removed, but I do think having an example output to help the reader understand what the package can do is helpful.
