### 2024-05-24 4338001
Update paper refs and acronyms
