+++ joss/paper.bib
@@ -45,3 +45,71 @@
   publisher={Nature Publishing Group},
   doi={10.1038/s41592-019-0686-2}
 }
+
+@book{Saltelli2007,
+author = {Saltelli, Andrea and Ratto, Marco and Andres, Terry and Campolongo, Francesca and Cariboni, Jessica and Gatelli, Debora and Saisana, Michaela and Tarantola, Stefano},
+booktitle = {Global Sensitivity Analysis. The Primer},
+doi = {10.1002/9780470725184},
+isbn = {9780470725184},
+month = {dec},
+pages = {237--275},
+publisher = {John Wiley {\&} Sons, Ltd},
+title = {{Global Sensitivity Analysis. The Primer}},
+year = {2007}
+}
+
+@article{sobol1993,
+  title={Sensitivity analysis for non-linear mathematical models, originally “Sensitivity estimates for non-linear mathematical models”},
+  author={Sobol, Ilya M},
+  journal={Math Model Comput Exp},
+  volume={1},
+  pages={407--414},
+  year={1993}
+}
+
+@report{europeancommission2021,
+  title = {Better {{Regulation Toolbox}}},
+  author = {{European Commission}},
+  date = {2021-11},
+  url = {https://ec.europa.eu/info/law/law-making-process/planning-and-proposing-law/better-regulation-why-and-how/better-regulation-guidelines-and-toolbox_en},
+  city = {Brussels},
+  keywords = {\#nosource}
+}
+
+@software{panel,
+  author       = {Philipp Rudiger and
+                  Marc Skov Madsen and
+                  Simon Høxbro Hansen and
+                  Maxime Liquet and
+                  Andrew and
+                  Xavier Artusi and
+                  James A. Bednar and
+                  Chris B and
+                  Jean-Luc Stevens and
+                  Christoph Deil and
+                  Demetris Roumis and
+                  Julia Signell and
+                  Mateusz Paprocki and
+                  Jerry Wu and
+                  Jon Mease and
+                  Arne and
+                  Coderambling and
+                  Hugues-Yanis Amanieu and
+                  thuydotm and
+                  Simon and
+                  sdc50 and
+                  Luca Fabbri and
+                  kbowen and
+                  Theom and
+                  Joel Ostblom and
+                  Govinda Totla and
+                  Niko Föhr and
+                  TBym},
+  title        = {holoviz/panel: Version 1.4.3},
+  month        = may,
+  year         = 2024,
+  publisher    = {Zenodo},
+  version      = {v1.4.3},
+  doi          = {10.5281/zenodo.11261266},
+  url          = {https://doi.org/10.5281/zenodo.11261266}
+}
