### 2024-05-30 c3ad212
Apply suggestions from code review
Co-authored-by: gnopik <37065157+gnopik@users.noreply.github.com>
