+++ joss/paper.md
@@ -51,8 +51,7 @@ Simulation Decomposition or SimDec moves the field of SA forward by supplementin
 In short, SimDec is a hybrid uncertainty-sensitivity analysis approach
 that reveals the critical behavior of a computational model or an empirical
 dataset. It decomposes the distribution of the output
-(target variable) by the multivariable scenarios, formed out of the most
-influential input variables. The resulting visualization shows how different
+(target variable) by automatically forming scenarios that reveal the most critical behavior of the system. The scenarios are formed out of the most influential input variables (defined with variance-based sensitivity indices) by breaking down their numeric ranges into states (e.g. _low_ and _high_) and creating an exhaustive list of their combinations (e.g. (i) _low_ _**A**_ & _low_ _**B**_, (ii) _low_ _**A**_ & _high_ _**B**_, (iii) _high_ **_A_** & _low_ **_B_**, and (iv) _high_ **_A_** and _high_ **_B_**). The resulting visualization shows how different
 output ranges can be achieved and what kind of critical interactions affect
 the output–as seen in \autoref{fig:simdec}. The method has shown value for
 various computational models from different fields, including business,
