### 2024-05-30 Simulation-Decomposition/simdec-python#35 @tupui
@matt-graham we addressed your points, thanks again. What remains is providing a script to reproduce the figure itself. Moving this to the other issues.
