### 2024-05-30 Simulation-Decomposition/simdec-python#31 @tupui
Closing in favour of #34
