### 2024-05-30 Simulation-Decomposition/simdec-python#31 @matt-graham
Just to note that having an example usage in the _documentation_ (not the JOSS paper) is a specific requirement JOSS reviewers are asked to check in their checklist

> Example usage: Do the authors include examples of how to use the software (ideally to solve real-world analysis problems).

If the intention of closing this in favour of #34 is that you'll still include an example usage in the documentation (of reproducing the figure from JOSS paper) that's fine but otherwise I would say this should be reopened.
