### 2024-05-30 6b27102
Add executable notebook in doc
