+++ environment.yml
@@ -22,3 +22,4 @@ dependencies:
   - pydata-sphinx-theme
   - accessible-pygments
   - numpydoc
+  - myst-nb
