### 2024-05-30 Simulation-Decomposition/simdec-python#34 @tupui
@matt-graham I added an executable notebook in the doc, see #23
