### 2024-05-30 Simulation-Decomposition/simdec-python#35 @matt-graham
Thanks @tupui, changes looks good. I've added a couple of small suggested changes to the updated text on #23.
