### 2024-05-30 Simulation-Decomposition/simdec-python#34 @matt-graham
Thanks this looks great.
