### 2024-05-30 dccfe3a
Apply suggestions from code review
Co-authored-by: Matt Graham <matthew.m.graham@gmail.com>
