### 2024-06-04 #6713 @crvernon
👋  @tupui  - we are almost there!

Please fix the following:

- [ ] LINE 56: This citation should have the "p" in Python capitalized.   You can retain capitalization in the bib file using curly brackets around the parts you want retained.

So here is what we have left to do:

- [ ]  Check the archival deposit (e.g., in Zenodo) to ensure it has the correct metadata. This includes the title (should match the paper title) and author list (make sure the list is correct and matches the paper, including the order they are placed). You may also add the authors' ORCID.

- [ ]  Please respond with the new version number and the DOI of the archived version here
