### 2024-06-05 6bb66ac
Fix capitalization in citations
