+++ joss/paper.bib
@@ -6,7 +6,7 @@
   volume = {2},
   number = {9},
   author = {Jon Herman and Will Usher},
-  title = {{SALib}: An open-source Python library for Sensitivity Analysis},
+  title = {{SALib}: An open-source {Python} library for Sensitivity Analysis},
   journal = {The Journal of Open Source Software}
 }
 
@@ -30,7 +30,7 @@
   number = {84},
   pages = {5309},
   author = {Pamphile T. Roy and Art B. Owen and Maximilian Balandat and Matt Haberland},
-  title = {Quasi-Monte Carlo Methods in Python},
+  title = {Quasi-Monte Carlo Methods in {Python}},
   journal = {Journal of Open Source Software}
 }
 
