### 2024-06-05 bad37d0
JOSS paper (#23)
Doc for JOSS:
https://joss.readthedocs.io/en/latest/submitting.html#example-paper-and-bibliography

---------

Co-authored-by: gnopik <37065157+gnopik@users.noreply.github.com>
Co-authored-by: Matt Graham <matthew.m.graham@gmail.com>
