+++ joss/paper.bib
@@ -1,0 +1,115 @@
+@article{Herman2017,
+  doi = {10.21105/joss.00097},
+  year  = {2017},
+  month = {jan},
+  publisher = {The Open Journal},
+  volume = {2},
+  number = {9},
+  author = {Jon Herman and Will Usher},
+  title = {{SALib}: An open-source {Python} library for Sensitivity Analysis},
+  journal = {The Journal of Open Source Software}
+}
+
+@article{Kozlova2024,
+  title = {Uncovering heterogeneous effects in computational models for sustainable decision-making},
+  journal = {Environmental Modelling & Software},
+  volume = {171},
+  pages = {105898},
+  year = {2024},
+  issn = {1364-8152},
+  doi = {10.1016/j.envsoft.2023.105898},
+  author = {Mariia Kozlova and Robert J. Moss and Julian Scott Yeomans and Jef Caers},
+  keywords = {Global sensitivity analysis, Simulation decomposition, Monte Carlo simulation, Decision-making under uncertainty}
+}
+
+@article{Roy2023,
+  doi = {10.21105/joss.05309},
+  year = {2023},
+  publisher = {The Open Journal},
+  volume = {8},
+  number = {84},
+  pages = {5309},
+  author = {Pamphile T. Roy and Art B. Owen and Maximilian Balandat and Matt Haberland},
+  title = {Quasi-Monte Carlo Methods in {Python}},
+  journal = {Journal of Open Source Software}
+}
+
+@article{Virtanen2020,
+  title={SciPy 1.0: fundamental algorithms for scientific computing in {P}ython},
+  author={Virtanen, Pauli and Gommers, Ralf and Oliphant, Travis E and Haberland, Matt and Reddy, Tyler and Cournapeau, David and Burovski, Evgeni and Peterson, Pearu and Weckesser, Warren and Bright, Jonathan et al.},
+  journal={Nature methods},
+  volume={17},
+  number={3},
+  pages={261--272},
+  year={2020},
+  publisher={Nature Publishing Group},
+  doi={10.1038/s41592-019-0686-2}
+}
+
+@book{Saltelli2007,
+author = {Saltelli, Andrea and Ratto, Marco and Andres, Terry and Campolongo, Francesca and Cariboni, Jessica and Gatelli, Debora and Saisana, Michaela and Tarantola, Stefano},
+booktitle = {Global Sensitivity Analysis. The Primer},
+doi = {10.1002/9780470725184},
+isbn = {9780470725184},
+month = {dec},
+pages = {237--275},
+publisher = {John Wiley {\&} Sons, Ltd},
+title = {{Global Sensitivity Analysis. The Primer}},
+year = {2007}
+}
+
+@article{sobol1993,
+  title={Sensitivity analysis for non-linear mathematical models, originally “Sensitivity estimates for non-linear mathematical models”},
+  author={Sobol, Ilya M},
+  journal={Math Model Comput Exp},
+  volume={1},
+  pages={407--414},
+  year={1993}
+}
+
+@report{europeancommission2021,
+  title = {Better {{Regulation Toolbox}}},
+  author = {{European Commission}},
+  date = {2021-11},
+  url = {https://ec.europa.eu/info/law/law-making-process/planning-and-proposing-law/better-regulation-why-and-how/better-regulation-guidelines-and-toolbox_en},
+  city = {Brussels},
+  keywords = {\#nosource}
+}
+
+@software{panel,
+  author       = {Philipp Rudiger and
+                  Marc Skov Madsen and
+                  Simon Høxbro Hansen and
+                  Maxime Liquet and
+                  Andrew and
+                  Xavier Artusi and
+                  James A. Bednar and
+                  Chris B and
+                  Jean-Luc Stevens and
+                  Christoph Deil and
+                  Demetris Roumis and
[…]
