+++ docs/tutorials.rst
@@ -1,0 +1,7 @@
+Executable tutorials
+====================
+
+.. toctree::
+   :maxdepth: 1
+
+   notebooks/structural_reliability
