### 2024-06-05 a44dd86
Remove JOSS workflow and prepare for next release
