### 2024-06-05 #6713 @tupui
Thank you @crvernon

- I fixed the capitalisation;
- I cut a release: https://github.com/Simulation-Decomposition/simdec-python/releases/tag/v1.2.0 ;
- The zenodo URL is: https://doi.org/10.5281/zenodo.11479751 (being processed).
