### 2024-06-05 #6713 @crvernon
@tupui it says permission required.  Can you set the record to public please?  Thanks
