### 2024-06-05 #6713 @tupui
It's not public yet, the article is still being process on their side. The visibility is set to public already.

Can you try this link? It's supposed to be a sharing link to view any resources https://zenodo.org/uploads/11479751?token=eyJhbGciOiJIUzUxMiJ9.eyJpZCI6IjcwMDUxYTA2LTk1ZGQtNGIwZS04NTlhLWFjZDVjNGUzYTcyNSIsImRhdGEiOnt9LCJyYW5kb20iOiI1NzE3MGMzYTE1MDljOTVmOWNlMDU3YjhkOGUwYjc3YSJ9.RWyGDkqX0_tBmG0blU34vO1R6gkWY3eqdIKMWf_seLJnnN0mokNWt5C-_wojqZ47sScrj1ETKbQ5fsBcLLaKYg
