### 2024-06-06 #6713 @crvernon
That link states that permission is denied.  I have never seen a Zenodo archive take so long to process.  We need the archive to be accessible to all before we can move forward.  Did you use Zenodo's web hook integration with GitHub to do generate the archive when you made the GitHub release?
