### 2024-06-06 #6713 @tupui
I connected with my GitHub auth but filled everything manually. I did not know there was a special thing for software. Shall I start over?
