### 2024-06-06 #6713 @crvernon
Yeah, something is off I believe. Usually I do the following workflow:

- Log in to Zenodo using your GitHub account
- In the top right hand corner of the screen where your login name is, click the dropdown and navigate to the "GitHub" tab
- If you don't see any repos, click  "Connect".  You should then see all GitHub repos you are admin on.
- Navigate to the one you are interested in and click the on/off toggle to on.
- Go to the target repo on GitHub and create the release.  This should automatically trigger an archive to start being formed on Zenodo.
- Navigate back to Zenodo and click on your repo of interest, the one you just toggled to "on"
- Once the archive says "Published" with a green check, click the DOI link to navigate to that record.

If you give me the link from that last step, I can run the rest from my side.  Thanks!
