### 2024-06-09 #6713 @tupui
Hi @crvernon, I deleted the draft on Zenodo and followed your instructions now to create a new version.

Was indeed immediate like that. (Probably a good idea to tell people to do this in the template.)

https://zenodo.org/records/11535796
